Fix Build Error (#183)
* schema.prisma定義、user,communityProfileと道案内関係は未定義

* Database Schema outline completed.

* Removed community-based modeling from db schema.

* DB: shortened ID column (hashedID) for URL routing

* DB: added details to schema

* installed package: zod, zod-prisma-types

* DB: zod-prisma-type settings

* Page: generateStaticParams moved to correct place
(layout.tsx -> page.tsx)

---------
